refactor: rename page files and components
