+++ webapp/client/apps/orchestration-cluster-webapp/src/pages/ForbiddenPage.tsx
@@ -6,28 +6,22 @@
  * except in compliance with the Camunda License 1.0.
  */
 
-import {Layer, Stack, Button} from '@carbon/react';
-import {Link} from '@tanstack/react-router';
+import {Layer, Stack} from '@carbon/react';
 import {useTranslation} from 'react-i18next';
-import SvgErrorRobot from '#/modules/svg/ErrorRobot';
-import styles from './notFound.module.scss';
+import SvgForbidden from '#/modules/svg/Forbidden';
+import styles from './ForbiddenPage.module.scss';
 
-const NotFound: React.FC = () => {
+const ForbiddenPage: React.FC = () => {
 	const {t} = useTranslation();
 
 	return (
 		<div className={styles['page']}>
 			<Layer withBackground className={styles['card']}>
 				<Stack orientation="horizontal" gap={6}>
-					<SvgErrorRobot aria-hidden />
-					<Stack gap={6}>
-						<Stack gap={3}>
-							<h1 className={styles['heading']}>{t('notFoundPageTitle')}</h1>
-							<p className={styles['description']}>{t('notFoundPageDescription')}</p>
-						</Stack>
-						<Button as={Link} to="/">
-							{t('notFoundPageButtonLabel')}
-						</Button>
+					<SvgForbidden aria-hidden />
+					<Stack gap={3}>
+						<h1 className={styles['heading']}>{t('forbiddenPageTitle')}</h1>
+						<p className={styles['description']}>{t('forbiddenPageDesc')}</p>
 					</Stack>
 				</Stack>
 			</Layer>
@@ -35,4 +29,4 @@ const NotFound: React.FC = () => {
 	);
 };
 
-export {NotFound};
+export {ForbiddenPage};
