+++ webapp/client/apps/orchestration-cluster-webapp/src/pages/GenericErrorPage.tsx
@@ -6,22 +6,27 @@
  * except in compliance with the Camunda License 1.0.
  */
 
-import {Layer, Stack} from '@carbon/react';
+import {Layer, Stack, Button} from '@carbon/react';
 import {useTranslation} from 'react-i18next';
-import SvgForbidden from '#/modules/svg/Forbidden';
-import styles from './forbidden.module.scss';
+import SvgErrorRobot from '#/modules/svg/ErrorRobot';
+import styles from './GenericErrorPagePage.module.scss';
 
-const Forbidden: React.FC = () => {
+type Props = {reset: () => void};
+
+const GenericErrorPage: React.FC<Props> = ({reset}) => {
 	const {t} = useTranslation();
 
 	return (
 		<div className={styles['page']}>
 			<Layer withBackground className={styles['card']}>
 				<Stack orientation="horizontal" gap={6}>
-					<SvgForbidden aria-hidden />
-					<Stack gap={3}>
-						<h1 className={styles['heading']}>{t('forbiddenPageTitle')}</h1>
-						<p className={styles['description']}>{t('forbiddenPageDesc')}</p>
+					<SvgErrorRobot aria-hidden />
+					<Stack gap={6}>
+						<Stack gap={3}>
+							<h1 className={styles['heading']}>{t('errorGenericErrorPageTitle')}</h1>
+							<p className={styles['description']}>{t('errorGenericErrorPageMessage')}</p>
+						</Stack>
+						<Button onClick={reset}>{t('errorGenericErrorPageButtonLabel')}</Button>
 					</Stack>
 				</Stack>
 			</Layer>
@@ -29,4 +34,4 @@ const Forbidden: React.FC = () => {
 	);
 };
 
-export {Forbidden};
+export {GenericErrorPage};
