+++ webapp/client/apps/orchestration-cluster-webapp/src/routes/_auth/admin/route.tsx
@@ -9,7 +9,7 @@
 import {createFileRoute, Outlet} from '@tanstack/react-router';
 import {getClientConfig} from '#/modules/config/getClientConfig';
 import {ComponentNotAvailableError} from '#/modules/errors/errors';
-import {Forbidden} from '#/pages/forbidden';
+import {ForbiddenPage} from '#/pages/ForbiddenPage';
 
 export const Route = createFileRoute('/_auth/admin')({
 	beforeLoad: () => {
@@ -19,7 +19,7 @@ export const Route = createFileRoute('/_auth/admin')({
 	},
 	errorComponent: ({error}) => {
 		if (error instanceof ComponentNotAvailableError) {
-			return <Forbidden />;
+			return <ForbiddenPage />;
 		}
 		throw error;
 	},
